Use `final` where possible
The `final` keyword in Java tells the reader (as well as the compiler)
that the parameter/field/variable is not going to have any other value
assigned to it after this statement. If we try and reassign a variable
it won't compile.

E.g:

final String s = "hello";
s = "world";

will not compile because we've changed the final variable 's'